backend/requirements.txt
absl-py==0.11.0
aiofile==3.3.3
aiofiles==0.6.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.2.1
caio==0.6.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
fastapi==0.62.0
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h11==0.11.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.3.21
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
uuid==1.30
uvicorn==0.13.2
virtualenv==20.2.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
